% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garnet_benchmark_01.R
\name{garnet_benchmark_01}
\alias{garnet_benchmark_01}
\title{garnet_benchmark_01}
\usage{
garnet_benchmark_01()
}
\value{
a database of a single XXX variety of garnet with the specimen(single mineral), sample, element and value (weight\%).
}
\description{
garnet_benchmark_01
}
\examples{
garnet_benchmark_01()
}
