% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garnet_endmember_fe2.R
\name{garnet_endmember_fe2}
\alias{garnet_endmember_fe2}
\title{garnet_endmember_fe2}
\usage{
garnet_endmember_fe2(database)
}
\arguments{
\item{database}{a dataframe with element atomic units per crystallographic site or a dataframe from garnet_site_occupancy()}
}
\value{
a dataframe with garnet endmembers proportion using the Fe3 and Fe2 for grossularia, pyrope, almandine, and spessartine - without the presence of spessartine, uvarovite, andradite and Ca-Ti garnets
}
\description{
garnet_endmember_fe2
}
\examples{
#' garnet_atom_units(garnet_benchmark_01()) \%>\%  # convert the weight\% chemistry data to atomic units
garnet_site_occupancy(.) \%>\%  # define site occupancy per element
garnet_endmember_fe2(.)
}
