% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garnet_endmember_fe3.R
\name{garnet_endmember_fe3}
\alias{garnet_endmember_fe3}
\title{garnet_endmember_fe3}
\usage{
garnet_endmember_fe3(database)
}
\arguments{
\item{database}{a dataframe with element atomic units per crystallographic site or a dataframe from garnet_site_occupancy()}
}
\value{
a dataframe with garnet endmembers proportion using the Fe+3 for spessartine, uvarovite, andradite and Ca-Ti garnets, without the presence of grossularia
}
\description{
garnet_endmember_fe3
}
\examples{
garnet_atom_units(garnet_benchmark_01()) \%>\%  # convert the weight\% chemistry data to atomic units
garnet_site_occupancy(.) \%>\%  # define site occupancy per element
garnet_endmember_fe3(.)

}
