% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garnet_EW.R
\name{garnet_EW}
\alias{garnet_EW}
\title{garnet_EW}
\usage{
garnet_EW()
}
\value{
a tibble of the elements, atomic weight of the cations, number of oxigens, and crystallographic site for each element of the mineral garnet.
}
\description{
garnet_EW
}
\examples{
garnet_EW()
}
