% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garnet_site_occupancy.R
\name{garnet_site_occupancy}
\alias{garnet_site_occupancy}
\title{garnet_site_occupancy}
\usage{
garnet_site_occupancy(database)
}
\arguments{
\item{database}{a dataframe from garnet_atom_units() function}
}
\value{
a dataframe wit
}
\description{
garnet_site_occupancy
}
\examples{
garnet_atom_units(garnet_benchmark_01(),garnet_EW()) \%>\%
garnet_site_occupancy(.)
}
