% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_to_percent.R
\name{ppm_to_percent}
\alias{ppm_to_percent}
\title{Convert PPM to Percentage}
\usage{
ppm_to_percent(ppm_value)
}
\arguments{
\item{ppm_value}{Numeric value representing the concentration in parts per million.}
}
\value{
A numeric value representing the concentration in percentage.
}
\description{
This function converts a value from parts per million (PPM) to a percentage.
}
\examples{
ppm_to_percent(10000) # returns 1
ppm_to_percent(5000)  # returns 0.5
}
