% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_concordia.R
\docType{data}
\name{StatConcordia}
\alias{StatConcordia}
\title{Concordia Statistical Transformation}
\format{
A ggproto object.
}
\usage{
StatConcordia
}
\description{
A ggproto object for computing concordia ellipses based on input data's covariance structure.
This transformation generates ellipse points centered at each (x, y) coordinate with the
specified covariance values, using \code{sigma_x}, \code{sigma_y}, and \code{rho} as inputs.
}
\keyword{datasets}
