% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WR_calculator.R
\name{WR_calculator}
\alias{WR_calculator}
\title{Whole-Rock (WR) Elemental Calculator}
\usage{
WR_calculator(
  BD,
  element_vector = c(y89, nb93, la139:lu175),
  database = "chapman_etal_2016.csv",
  element_plus_size = TRUE,
  error_tag = "_2s",
  normalized = TRUE,
  tag = ""
)
}
\arguments{
\item{BD}{A data frame containing elemental concentration data.}

\item{element_vector}{A vector of elements to include for WR normalization (default: \code{c(y89, nb93, la139:lu175)}).}

\item{database}{The name of the CSV file containing reference values (\code{"chapman_etal_2016.csv"}).}

\item{element_plus_size}{Logical indicating if elements with size should be used from \code{database} (default is TRUE).}

\item{error_tag}{The suffix to ignore in \code{BD} when selecting columns (default is \code{"_2s"}).}

\item{normalized}{Logical indicating if the result should be normalized (default is TRUE).}

\item{tag}{Optional tag to append to elements in the output (default is "").}
}
\value{
A data frame with WR normalized values for each element, with suffix \verb{_WR}.
}
\description{
This function calculates whole-rock (WR) normalized values for selected elements in a given dataset.
It uses a reference dataset based on Chapman et al., 2016 for normalization. The Chapman et al. dataset provides
partition coeficients to calculate WR elements based on Zircon trace element data.
}
\details{
Parent-rock concentrations are estimated as \code{C_zircon / (a * C_zircon^b)} with the
coefficients of Chapman, J.B., Gehrels, G.E., Ducea, M.N., Giesler, N. and Pullen, A., 2016.
A new method for estimating parent rock trace element concentrations from zircon.
\emph{Chemical Geology}, 439, 59-70.
}
