% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WR_crustal_thickness.R
\name{WR_crustal_thickness}
\alias{WR_crustal_thickness}
\title{Calculate Whole-Rock Crustal Thickness}
\usage{
WR_crustal_thickness(BD)
}
\arguments{
\item{BD}{A data frame containing the trace element data.
It must include columns \code{la139_WR_N} and \code{yb172_WR_N} for normalized La and Yb values, respectively.}
}
\value{
A data frame with an additional column \code{WR_crust} representing the calculated whole-rock crustal thickness.
}
\description{
This function estimates the whole-rock crustal thickness based on La/Yb normalized trace element ratios in zircon.
The data and method are based on the work of Profeta, L., Ducea, M.N., Chapman, J.B., Paterson, S.R., Gonzales, S.M.H., Kirsch, M., Petrescu, L. and DeCelles, P.G., 2015. Quantifying crustal thickness over time in magmatic arcs. Scientific reports, 5(1), p.17786.
}
\examples{
\dontrun{
  BD <- data.frame(la139_WR_N = c(0.1, 0.2), yb172_WR_N = c(0.05, 0.1))
  WR_crustal_thickness(BD)
}
}
\references{
Chapman, J.B., Gehrels, G.E., Ducea, M.N., Giesler, N. and Pullen, A. (2016). A new method for estimating parent rock trace element concentrations from zircon. \emph{Chemical Geology}, 439, pp.59-70.
}
