% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zircon_ti_f_site.R
\name{zircon_ti_f_site}
\alias{zircon_ti_f_site}
\title{Calculate the Fraction of Ti on the Si-Site in Zircon}
\usage{
zircon_ti_f_site(P)
}
\arguments{
\item{P}{Numeric. Pressure in gigapascals (GPa).}
}
\value{
Numeric. The fraction of Ti on the Si-site in zircon.
}
\description{
This function calculates the fraction of titanium (Ti) on the Si-site in zircon as a function of pressure.
}
\examples{
# Example usage:
zircon_ti_f_site(1.5)


}
