% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_formula.R
\name{chemical_formula}
\alias{chemical_formula}
\title{Calculate Atoms Per Formula Unit (APFU) for Minerals}
\usage{
chemical_formula(
  dataframe,
  oxigens,
  cations,
  weight_location = "/data/element_weights.csv",
  base = "anions"
)
}
\arguments{
\item{dataframe}{DataFrame. A dataframe with element oxides in \%weight.}

\item{oxigens}{Numeric. Number of oxygens used for base calculation.}

\item{cations}{Numeric. Number of cations used for base calculation.}

\item{weight_location}{String. The file location of an in-house oxides weight dataframe, otherwise will use the one in the folder ./data.}

\item{base}{String. Either "anions" or "cations" to indicate the basis of the calculation.}
}
\value{
DataFrame with atoms per formula unit (APFU) for each element without crystallographic site balancing.
}
\description{
This function calculates the atoms per formula unit (APFU) for each element in a given mineral dataset.
The calculations can be based on either the number of oxygens (anions) or cations.
}
\examples{
# Example for Tourmaline
read.csv("./data/tourmaline_DHZ.csv") \%>\%
  chemical_formula(oxigens = 31, cations = 18, base = "anions")

# Example for Epidote
read.csv("./data/epidote_DHZ.csv") \%>\%
  chemical_formula(oxigens = c(12.5, 13, 12.5, 12.5, 12.5, 13), cations = 8, base = "cations")
}
