% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_isoplotr.R
\name{use_isoplotr}
\alias{use_isoplotr}
\title{Calculate U-Pb Ages Using IsoplotR}
\usage{
use_isoplotr(df, age_type = 1)
}
\arguments{
\item{df}{A data frame with columns for isotopic ratios and uncertainties (\code{pb207_u235}, \code{pb207_u235_2s},
\code{pb206_u238}, \code{pb206_u238_2s}, \code{pb207_pb206}, \code{pb207_pb206_2s}, \code{rho_207pb_206pb_v_238u_206pb}).}

\item{age_type}{Numeric specifying the type of age to compute in \code{IsoplotR::age}. Default is 1.}
}
\value{
A data frame combining the input \code{df} with calculated ages (\code{age_75}, \code{age_68}, \code{age_76}, and concordia age).
}
\description{
This function calculates U-Pb ages using \code{IsoplotR::age}, handling possible errors and returning
a data frame with age values and associated uncertainties.
}
\examples{
# Assuming `df` is a data frame with the required columns
use_isoplotr(df, age_type = 1)
}
