% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zircon_ti_p.R
\name{zircon_ti_p}
\alias{zircon_ti_p}
\title{Calculate the Pressure for a Given Ti Concentration in Zircon}
\usage{
zircon_ti_p(ti)
}
\arguments{
\item{ti}{Numeric. The value of Ti concentration in ppm.}
}
\value{
Numeric. The pressure in gigapascals (GPa) or NA if no real solution exists.
}
\description{
This function calculates the pressure (in GPa) for a given Ti concentration in zircon (Ti concentration in ppm).
The function uses a quadratic equation to solve for pressure.
}
\examples{
# Example usage:
zircon_ti_p(3.0)


}
