% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zircon_ti_t.R
\name{zircon_ti_t}
\alias{zircon_ti_t}
\title{Calculate zircon crystallization temperature}
\usage{
zircon_ti_t(pressure, aSiO2 = 1, aTiO2 = 1, ti_ppm, equation = "watson")
}
\arguments{
\item{pressure}{Pressure in giga Pascal.}

\item{aSiO2}{Activity fraction of SiO2 (default: 1).}

\item{aTiO2}{Activity fraction of TiO2 (default: 1).}

\item{ti_ppm}{Concentration of titanium in parts per million.}

\item{equation}{Equation to use for calculation: "crisp" or "watson" (default: "crisp").}
}
\value{
Crystallization temperature of zircon in Celsius.
}
\description{
This function calculates the crystallization temperature of zircon based on various parameters.
}
\details{
References
Crisp, Laura J., Andrew J. Berry, Antony D. Burnham, Laura A. Miller, and Matthew Newville. 2023. “The Ti-in-Zircon Thermometer Revised: The Effect of Pressure on the Ti Site in Zircon.” Geochimica Et Cosmochimica Acta 360 (November): 241–58. https://doi.org/10.1016/j.gca.2023.04.031.
Watson, E. B., D. A. Wark, and J. B. Thomas. 2006. “Crystallization Thermometers for Zircon and Rutile.” Contributions to Mineralogy and Petrology 151 (4): 413–33. https://doi.org/10.1007/s00410-006-0068-5.
}
\examples{
zircon_ti_t(pressure = 5, ti_ppm = 50)

}
